+++ spec/tests/ConvivaAnalyticsTracker.spec.ts
@@ -2,7 +2,6 @@ import { PlayerEvent, PlayerEventBase } from "bitmovin-player";
 import { ConvivaAnalyticsTracker } from "../../src/ts/ConvivaAnalyticsTracker";
 import { MockHelper } from "../helper/MockHelper";
 import * as Conviva from '@convivainc/conviva-js-coresdk';
-import { PlayerStateHelper } from "../../src/ts/helper/PlayerStateHelper";
 
 jest.mock('@convivainc/conviva-js-coresdk', () => {
   const { MockHelper } = jest.requireActual('../helper/MockHelper');
@@ -78,6 +77,40 @@ describe(ConvivaAnalyticsTracker, () => {
     expect(invokedTimesAfter).toBe(invokedTimesBefore);
   })
 
+  it('should report ad break ended on RestoringContent event', () => {
+    const { playerMock, playerEventHelper } = MockHelper.createPlayerMock();
+    const convivaAnalyticsTracker = new ConvivaAnalyticsTracker('test-key');
+
+    convivaAnalyticsTracker.attachPlayer(playerMock);
+    playerEventHelper.firePlayEvent();
+    convivaAnalyticsTracker.trackRestoringContent();
+
+    expect(MockHelper.latestVideoAnalytics.reportAdBreakEnded).toHaveBeenCalledTimes(1);
+  });
+
+  it('should not report the player state on AdBreakFinished events if there is still an active ad', () => {
+    const { playerMock, playerEventHelper } = MockHelper.createPlayerMock();
+    const convivaAnalyticsTracker = new ConvivaAnalyticsTracker('test-key');
+
+    convivaAnalyticsTracker.attachPlayer(playerMock);
+    playerEventHelper.fireAdBreakFinishedEvent();
+    convivaAnalyticsTracker.trackAdBreakFinished();
+
+    expect(MockHelper.latestVideoAnalytics.reportPlaybackMetric).not.toHaveBeenCalled();
+  });
+
+  it('should report the player state on AdBreakFinished events if there is no active ad', () => {
+    const { playerMock, playerEventHelper } = MockHelper.createPlayerMock();
+    const convivaAnalyticsTracker = new ConvivaAnalyticsTracker('test-key');
+
+    convivaAnalyticsTracker.attachPlayer(playerMock);
+    playerEventHelper.firePlayEvent();
+    convivaAnalyticsTracker.trackRestoringContent();
+    convivaAnalyticsTracker.trackAdBreakFinished();
+
+    expect(MockHelper.latestVideoAnalytics.reportPlaybackMetric).toHaveBeenCalled();
+  });
+
   describe('trackPlaybackStateChanged', () => {
     let convivaAnalyticsTracker: ConvivaAnalyticsTracker;
 
@@ -113,7 +146,6 @@ describe(ConvivaAnalyticsTracker, () => {
       PlayerEvent.StallEnded,
       PlayerEvent.PlaybackFinished,
       PlayerEvent.AdStarted,
-      PlayerEvent.AdBreakFinished,
     ])('should clear timer for stalling when reported player event is %s', (event) => {
       const stallTrackingStopTimeoutSpy = jest.spyOn(convivaAnalyticsTracker['stallTrackingTimeout'], 'clear');
 
@@ -122,117 +154,6 @@ describe(ConvivaAnalyticsTracker, () => {
       expect(stallTrackingStopTimeoutSpy).toHaveBeenCalled();
     })
   })
-
-  /**
-   * Since the web player dispatches AdBreakFinished only after main content has successfully restored,
-   * a workaround is in place to signal the end of the ad break early so that VST can properly be tracked.
-   */
-  describe('eager ad break ended signalling', () => {
-    let convivaAnalyticsTracker: ConvivaAnalyticsTracker;
-    let reportAdBreakEndedSpy: jest.SpyInstance;
-    let reportAdBreakStartedSpy: jest.SpyInstance;
-
-    beforeEach(() => {
-      convivaAnalyticsTracker = new ConvivaAnalyticsTracker('test-key');
-
-      const {playerMock} = MockHelper.createPlayerMock();
-      convivaAnalyticsTracker.attachPlayer(playerMock);
-      jest.spyOn(playerMock, 'getSource').mockImplementation(() => ({ title: 'test-title' }));
-      convivaAnalyticsTracker.initializeSession();
-      
-      jest.useFakeTimers();
-      reportAdBreakEndedSpy = jest.spyOn(convivaAnalyticsTracker['convivaVideoAnalytics'], 'reportAdBreakEnded');
-      reportAdBreakStartedSpy = jest.spyOn(convivaAnalyticsTracker['convivaVideoAnalytics'], 'reportAdBreakStarted');
-    });
-
-    it('should report ad break ended on AdBreakFinished by default', () => {
-      convivaAnalyticsTracker.trackAdBreakStarted(Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdFinished();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(0);
-
-      convivaAnalyticsTracker.trackAdBreakFinished();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-    })
-
-    it('should report ad break ended early if AdBreakFinished is too slow', () => {
-      convivaAnalyticsTracker.trackAdBreakStarted(Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdFinished();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(0);
-
-      jest.runAllTimers();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-    })
-
-    it('should not report ad break ended twice if AdBreakFinished is emitted after the ad finished timer has already expired', () => {
-      convivaAnalyticsTracker.trackAdBreakStarted(Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdFinished();
-
-      jest.runAllTimers();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-
-      convivaAnalyticsTracker.trackAdBreakFinished();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-    })
-
-    it('should not report ad break ended if AdFinished is followed by a new AdStarted', () => {
-      convivaAnalyticsTracker.trackAdBreakStarted(Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-
-      expect(reportAdBreakStartedSpy).toHaveBeenCalledTimes(1);
-
-      convivaAnalyticsTracker.trackAdFinished();
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-
-      jest.runAllTimers();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(0);
-      expect(reportAdBreakStartedSpy).toHaveBeenCalledTimes(1);
-    })
-
-    it('should report ad break started again after wrongly reporting ad break ended', () => {
-      convivaAnalyticsTracker.trackAdBreakStarted(Conviva.Constants.AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-
-      expect(reportAdBreakStartedSpy).toHaveBeenCalledTimes(1);
-
-      convivaAnalyticsTracker.trackAdFinished();
-      jest.runAllTimers();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-
-      convivaAnalyticsTracker.trackAdStarted({}, Conviva.Constants.AdType.CLIENT_SIDE);
-
-      expect(reportAdBreakStartedSpy).toHaveBeenCalledTimes(2);
-    })
-
-    it('should only signal that playback has resumed after the actual AdBreakFinished', () => {
-      const { AdType, Playback, PlayerState } = Conviva.Constants;
-      const reportPlaybackSpy = jest.spyOn(convivaAnalyticsTracker['convivaVideoAnalytics'], 'reportPlaybackMetric');
-      jest.spyOn(PlayerStateHelper, 'getPlayerState').mockReturnValue(PlayerState.PLAYING);
-
-      convivaAnalyticsTracker.trackAdBreakStarted(AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdStarted({}, AdType.CLIENT_SIDE);
-      convivaAnalyticsTracker.trackAdFinished();
-
-      jest.runAllTimers();
-
-      expect(reportAdBreakEndedSpy).toHaveBeenCalledTimes(1);
-      expect(reportPlaybackSpy).not.toHaveBeenCalledWith(Playback.PLAYER_STATE, PlayerState.PLAYING);
-
-      convivaAnalyticsTracker.trackAdBreakFinished();
-
-      expect(reportPlaybackSpy).toHaveBeenCalledWith(Playback.PLAYER_STATE, PlayerState.PLAYING);
-    })
-  });
 })
 
 const getInvokedTimes = (mock: unknown) => {